+++ source/games/exhumed/src/lavadude.cpp
@@ -40,7 +40,7 @@ static actionSeq LavadudeSeq[] = {
 DExhumedActor* BuildLavaLimb(DExhumedActor* pActor, int move, int ebx)
 {
     auto pSprite = &pActor->s();
-    short nSector = pSprite->sectnum;
+    int nSector =pSprite->sectnum;
 
     auto pLimbActor = insertActor(nSector, 118);
 	auto pLimbSprite = &pLimbActor->s();
@@ -104,7 +104,7 @@ void AILavaDudeLimb::Draw(RunListEvent* ev)
 }
 
 
-void BuildLava(DExhumedActor* pActor, int x, int y, int, short nSector, short nAngle, int nChannel)
+void BuildLava(DExhumedActor* pActor, int x, int y, int, int nSector, short nAngle, int nChannel)
 {
     spritetype* pSprite;
     if (pActor == nullptr)
@@ -296,7 +296,7 @@ void AILavaDude::Tick(RunListEvent* ev)
         int x = pSprite->x;
         int y = pSprite->y;
         int z = pSprite->z;
-        short nSector = pSprite->sectnum;
+        int nSector =pSprite->sectnum;
 
         auto coll = movesprite(pActor, pSprite->xvel << 8, pSprite->yvel << 8, 0, 0, 0, CLIPMASK0);
